Rename module logical-query-builder to logical-plan-builder.
Plus some other random changes.

+++ community/cypher/logical-plan-builder/src/main/scala/org/neo4j/cypher/internal/logical/builder/AbstractLogicalPlanBuilder.scala
@@ -126,7 +126,7 @@ abstract class AbstractLogicalPlanBuilder[T, IMPL <: AbstractLogicalPlanBuilder[
     self
   }
 
-  def limit(count: Int): IMPL = {
+  def limit(count: Long): IMPL = {
     appendAtCurrentIndent(UnaryOperator(lp => Limit(lp, literalInt(count), DoNotIncludeTies)(_)))
     self
   }
